Fix painting scoring bug

+++ src/moves.rs
@@ -23,7 +23,7 @@ pub enum Move {
 }
 
 #[derive(Debug, Clone, Copy, PartialEq, Add, AddAssign)]
-pub struct Cost(pub u32);
+pub struct Cost(pub u64);
 
 impl Display for Orientation {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
@@ -68,7 +68,7 @@ impl Move {
     }
 
     fn compute_cost(&self, block_area: u32, canvas_area: u32) -> Cost {
-        Cost((self.base_cost() as f32 * (canvas_area as f32 / block_area as f32)).round() as u32)
+        Cost((self.base_cost() as f32 * (canvas_area as f32 / block_area as f32)).round() as u64)
     }
 
     fn color(
